moved comparison tests to `SetAlgebraTests` case
